Changes to template as described in readme
The first page of each chapter shows the chapter number in a very big font above the title, similar to the classicthesis template.
Section titles are preceded by the section number.
Section titles have a bigger font and are not italized.
It supports Parts (e.g. Part I, Part II), which also appear in the table of contents.
The table of content shows the chapter and section numbers next to their titles.
The cover is very simple and doesn't show any university-specific information (for MIT's version, check Tiff's template).
The pages before the "main matter" (the main content of the thesis after the cover, table of contents, acknowledgments, etc) are numbered with arabic characters (i, ii, iii). The first page of the "main matter" uses numbers starting from 1.
It adds the trabstractpage command for adding a translated abstract, ideal for universities that require an additional abstract in the local language.